refactor(product): ProductList View(상품 리스트 정보 뷰) 찜 목록, 거래 라벨 추가
